Rename ValueRefParserImpl.cpp to ValueRefParser.cpp
ValueRefParserImpl.h is an empty file and ValueRefParserImpl.cpp only
implements things declared in ValueRefParser.h
