+++ msvc2015/Parsers/Parsers.vcxproj
@@ -189,7 +189,7 @@
     <ClCompile Include="..\..\parse\TechsParser.cpp" />
     <ClCompile Include="..\..\parse\Tokens.cpp" />
     <ClCompile Include="..\..\parse\UniverseObjectTypeValueRefParser.cpp" />
-    <ClCompile Include="..\..\parse\ValueRefParserImpl.cpp" />
+    <ClCompile Include="..\..\parse\ValueRefParser.cpp" />
     <ClCompile Include="..\..\parse\VisibilityValueRefParser.cpp" />
   </ItemGroup>
   <ItemGroup>
@@ -218,7 +218,6 @@
     <ClInclude Include="..\..\parse\ReportParseError.h" />
     <ClInclude Include="..\..\parse\Tokens.h" />
     <ClInclude Include="..\..\parse\ValueRefParser.h" />
-    <ClInclude Include="..\..\parse\ValueRefParserImpl.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
