+++ parse/CMakeLists.txt
@@ -26,7 +26,6 @@ target_sources(freeorionparseobj
         ${CMAKE_CURRENT_LIST_DIR}/ReportParseError.h
         ${CMAKE_CURRENT_LIST_DIR}/Tokens.h
         ${CMAKE_CURRENT_LIST_DIR}/ValueRefParser.h
-        ${CMAKE_CURRENT_LIST_DIR}/ValueRefParserImpl.h
         ${CMAKE_CURRENT_LIST_DIR}/BuildingsParser.cpp
         ${CMAKE_CURRENT_LIST_DIR}/CommonParamsParser.cpp
         ${CMAKE_CURRENT_LIST_DIR}/ConditionParser1.cpp
@@ -73,6 +72,6 @@ target_sources(freeorionparseobj
         ${CMAKE_CURRENT_LIST_DIR}/TechsParser.cpp
         ${CMAKE_CURRENT_LIST_DIR}/Tokens.cpp
         ${CMAKE_CURRENT_LIST_DIR}/UniverseObjectTypeValueRefParser.cpp
-        ${CMAKE_CURRENT_LIST_DIR}/ValueRefParserImpl.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/ValueRefParser.cpp
         ${CMAKE_CURRENT_LIST_DIR}/VisibilityValueRefParser.cpp
 )
