+++ parse/ConditionParser6.cpp
@@ -1,6 +1,6 @@
 #include "ConditionParser6.h"
 
-#include "ValueRefParserImpl.h"
+#include "ValueRefParser.h"
 
 #include "../universe/Condition.h"
 #include "../universe/ValueRef.h"
