+++ parse/ConditionParser7.cpp
@@ -1,6 +1,6 @@
 #include "ConditionParser7.h"
 
-#include "ValueRefParserImpl.h"
+#include "ValueRefParser.h"
 
 #include "../universe/Condition.h"
 #include "../universe/ValueRef.h"
