+++ parse/EffectParser2.cpp
@@ -1,6 +1,6 @@
 #include "EffectParser2.h"
 
-#include "ValueRefParserImpl.h"
+#include "ValueRefParser.h"
 #include "../universe/Effect.h"
 #include "../universe/Enums.h"
 #include "../universe/ValueRef.h"
