+++ parse/EffectParser3.cpp
@@ -1,6 +1,6 @@
 #include "EffectParser3.h"
 
-#include "ValueRefParserImpl.h"
+#include "ValueRefParser.h"
 #include "../universe/Effect.h"
 #include "../universe/ValueRef.h"
