+++ parse/IntValueRefParser.cpp
@@ -1,4 +1,4 @@
-#include "ValueRefParserImpl.h"
+#include "ValueRefParser.h"
 
 #include "MovableEnvelope.h"
 #include "../universe/ValueRef.h"
