+++ parse/SpeciesParser.cpp
@@ -2,7 +2,7 @@
 
 #include "ParseImpl.h"
 #include "EnumParser.h"
-#include "ValueRefParserImpl.h"
+#include "ValueRefParser.h"
 #include "EffectParser.h"
 #include "ConditionParserImpl.h"
 #include "MovableEnvelope.h"
